requirements.txt
Django==1.9.8
arrow==0.8.0
beautifulsoup4==4.5.0
django-debug-toolbar==1.5.0
django-disqus==0.5.0
django-emoticons==1.2.0
django-forms-bootstrap==3.0.1
django-paypal==0.3.2
django-tinymce==2.3.0
python-dateutil==2.5.3
stripe
git+https://github.com/PramathaMadhavankutty/reusable_blog_app.git
